made snippet.el part of smart-snippet.el, that is made smart-snippet.el a fork of snippet.el
And add a new feature: now you can escape the identifier with \
For this example from Philipp Grau

(require 'cperl-mode)
(smart-snippet-with-abbrev-tables
(cperl-mode-abbrev-table)
("fh" "foreach my $key (keys %$${HASH}) {\n
     $>$value = $$${HASH}{\$key};\n
     $>$. # do something \n
     }$>"
     'bol?) ;
     )

So there are two places were HASH should be changable, one with a % in
front and the second one whit a $ in front, and this is my
problem. I'm not able to get the $-sign and directly after that the
snippet-field.

now it can be defined by

(require 'cperl-mode)
(smart-snippet-with-abbrev-tables
(cperl-mode-abbrev-table)
("fh" "foreach my $key (keys %$${HASH}) {\n
     $>$value = \\$$${HASH}{\$key};\n
     $>$. # do something \n
     }$>"
     'bol?) ;
     )


git-svn-id: http://smart-snippet.googlecode.com/svn/trunk@28 18804ea3-c42f-0410-b318-4510d07514a1